Optimize run_sync & test cleanup (#171)
Changes:

- optimize run_sync
- cleanup and speedup tests+some fixes
- fix typings
- remove unused dependency (pytest-asyncio)

+++ tests/clauses/test_not_clauses.py
@@ -2,10 +2,11 @@ import pytest
 
 import edgy
 from edgy.core.db.querysets.clauses import not_
+from edgy.testclient import DatabaseTestClient
 from tests.settings import DATABASE_URL
 
-database = edgy.Database(url=DATABASE_URL)
-models = edgy.Registry(database=database)
+database = DatabaseTestClient(DATABASE_URL)
+models = edgy.Registry(database=edgy.Database(database, force_rollback=True))
 
 pytestmark = pytest.mark.anyio
 
@@ -19,12 +20,19 @@ class User(edgy.Model):
         registry = models
 
 
-@pytest.fixture(autouse=True, scope="function")
+@pytest.fixture(autouse=True, scope="module")
 async def create_test_database():
     async with database:
         await models.create_all()
         yield
-    await models.drop_all()
+        if not database.drop:
+            await models.drop_all()
+
+
+@pytest.fixture(autouse=True, scope="function")
+async def rollback_transactions():
+    async with models.database:
+        yield
 
 
 async def test_filter_with_not():
