+++ tests/foreign_keys/m2m_string/test_many_to_many_field.py
@@ -7,7 +7,7 @@ from tests.settings import DATABASE_URL
 
 pytestmark = pytest.mark.anyio
 
-database = DatabaseTestClient(DATABASE_URL)
+database = DatabaseTestClient(DATABASE_URL, full_isolation=False)
 models = edgy.Registry(database=database)
 
 
@@ -45,18 +45,13 @@ class Studio(edgy.Model):
         registry = models
 
 
-@pytest.fixture(autouse=True)
-async def create_test_database():
-    await models.create_all()
-    yield
-    await models.drop_all()
-
-
 @pytest.fixture(autouse=True, scope="function")
-async def rollback_connections():
-    with database.force_rollback():
-        async with database:
-            yield
+async def create_test_database():
+    async with database:
+        await models.create_all()
+        yield
+        if not database.drop:
+            await models.drop_all()
 
 
 async def test_add_many_to_many():
