+++ tests/integration/test_esmerald_fk_reference_middleware.py
@@ -11,27 +11,25 @@ from edgy.testclient import DatabaseTestClient
 from tests.settings import DATABASE_URL
 
 database = DatabaseTestClient(DATABASE_URL)
-models = edgy.Registry(database=database)
+models = edgy.Registry(database=edgy.Database(database, force_rollback=True))
 
 pytestmark = pytest.mark.anyio
 pydantic_version = __version__[:3]
 
 
-@pytest.fixture(autouse=True, scope="function")
+@pytest.fixture(autouse=True, scope="module")
 async def create_test_database():
-    try:
+    async with database:
         await models.create_all()
         yield
-        await models.drop_all()
-    except Exception:
-        pytest.skip("No database available")
+        if not database.drop:
+            await models.drop_all()
 
 
-@pytest.fixture(autouse=True)
-async def rollback_connections():
-    with database.force_rollback():
-        async with database:
-            yield
+@pytest.fixture(autouse=True, scope="function")
+async def rollback_transactions():
+    async with models.database:
+        yield
 
 
 def blocking_function():
