+++ tests/models/test_inner_select.py
@@ -8,13 +8,14 @@ from edgy.contrib.multi_tenancy import TenantRegistry
 from edgy.testclient import DatabaseTestClient
 from tests.settings import DATABASE_URL
 
-database = DatabaseTestClient(DATABASE_URL)
+database = DatabaseTestClient(DATABASE_URL, full_isolation=True)
 models = TenantRegistry(database=database)
 
 
 pytestmark = pytest.mark.anyio
 pydantic_version = __version__[:3]
 
+# TODO: disallow loading and check the crashes
 
 class EdgyTenantBaseModel(edgy.Model):
     id: int = edgy.IntegerField(primary_key=True)
@@ -67,18 +68,13 @@ class Permission(EdgyTenantBaseModel):
         tablename = "ut_permission"
 
 
-@pytest.fixture(autouse=True, scope="module")
+@pytest.fixture(autouse=True, scope="function")
 async def create_test_database():
-    await models.create_all()
-    yield
-    await models.drop_all()
-
-
-@pytest.fixture(autouse=True)
-async def rollback_connections():
-    with database.force_rollback():
-        async with database:
-            yield
+    async with database:
+        await models.create_all()
+        yield
+        if not database.drop:
+            await models.drop_all()
 
 
 async def test_inner_select():
